Fixed the actuator data type.

+++ lexxpluss_apps/src/adc_reader.hpp
@@ -5,7 +5,7 @@
 struct adc_reader {
     static void init();
     static void run(void *p1, void *p2, void *p3);
-    static uint16_t get(int index);
+    static int32_t get(int index);
     static k_thread thread;
     static constexpr int INDEX_DOWNWARD_L{0};
     static constexpr int INDEX_DOWNWARD_R{1};
